feat: refactor the test cases and config
